requirements.txt
Django==1.11.4
Fabric==1.8.0
Jinja2==2.7.2
MarkupSafe==0.18
Pillow==2.9.0
Pygments==1.6
Sphinx==1.2
coverage==4.4.1
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-extensions==1.5.9
docutils==0.11
ecdsa==0.10
freezegun==0.3.9
paramiko==1.12.0
pycrypto==2.6.1
python-dateutil==2.4.2
releases==0.2.4
six==1.10.0
sphinxcontrib-googleanalytics==0.1
sphinxcontrib-programoutput==0.8
sqlparse==0.1.18
